refactor: consolidate to standard Medallion architecture
- Moved FMP API client from ingestion/utils.py to bronze/utils.py
- Renamed fmp_bronze.py to fmp.py (standard naming)
- Removed redundant ingestion/ folder
- Updated all imports to use bronze.utils and bronze.ingestion
- Updated pyproject.toml package discovery
- Updated CI workflow import tests
- Follows clean Medallion: bronze/ → silver/ → validation/
