--- ingestion/__init__.py
@@ -1,6 +1,0 @@
-"""E2E Stock Data Pipeline v2 - Ingestion Module."""
-
-# Make key classes available at package level
-from .utils import AsyncFMPClient, FMPConfig, RateLimiter
-
-__all__ = ["AsyncFMPClient", "FMPConfig", "RateLimiter"]
